RLIB-27: Converge back to 2.0.x API, phase 6
Moved back from gint64 to gint in more internal APIs.

+++ libsrc/charencoder.c
@@ -51,7 +51,7 @@ void rlib_charencoder_free(GIConv converter UNUSED) {
 #endif
 }
 
-gint64 rlib_charencoder_convert(GIConv converter UNUSED, gchar **inbuf, gsize *inbytes_left UNUSED, gchar **outbuf, gsize *outbytes_left UNUSED) {
+gint rlib_charencoder_convert(GIConv converter UNUSED, gchar **inbuf, gsize *inbytes_left UNUSED, gchar **outbuf, gsize *outbytes_left UNUSED) {
 #ifdef DISABLE_UTF8
 	/* The strlen is passed in here so we bump it by 1 */
 	*outbuf = g_strdup(*inbuf);
